// emdb/emdb-cli/src/model/MovieCreateRequest.ts
import { ShowStatus } from "./ShowStatus.js";

export interface MovieCreateRequest {
  source: string;
  externalId: number;
  title: string;
  releaseDate: string | null;
  score: number | null;
  status: ShowStatus;
  runtime: number | null;
  budget: number | null;
  revenue: number | null;
  homepage: string | null;
  originalLanguage: string | null;
  backdrop: string | null;
  poster: string | null;
  tagline: string | null;
  overview: string | null;
}