// emdb/emdb-cli/src/model/MovieUpdateRequest.ts
import { ShowStatus } from "./ShowStatus.js";

export interface MovieUpdateRequest {
  title: string | null;
  releaseDate: string | null;
  score: number | null;
  status: ShowStatus | null;
  runtime: number | null;
  budget: number | null;
  revenue: number | null;
  homepage: string | null,
  originalLanguage: string | null;
  backdrop: string | null;
  poster: string | null;
  tagline: string | null;
  overview: string | null;
}